// frontend/src/components/user_components/common/service_card/type.ts
export interface serviceCardProps {
    id?: number;
    imgSrc?: string;
    image?: string;
    title?: string;
    service_name?: string;
    description?: string;
  }
  
export interface projectCardProps {
    imgSrc?: string;
    image?: string;
    title?: string;
    service_name?: string;
    description?: string;
  }
export interface bannerProps {
  title?: string;
  description?: string
  image?: string
  }
  